chore(storage): rename cloud storage modules for clarity
